MENU UPDATE 2
The menu assets should all be in place. The only issue is the main gameplay level wont allow me to place a background without it obscuring the middle triangle and the emotes. This is because of the way the prefab in the level works but even unpacking it didnt help. so if anyone knows a way around this I'd much appreciate the help.